Fix names of files and folders
