+++ src/renderer/components/CustomRouter/CustomRouter.scss
@@ -1,0 +1,4 @@
+.router {
+    display: flex;
+    height: 100%;
+}
