+++ src/renderer/components/NavigationBar/NavigationBar.scss
