+++ src/renderer/components/RouterComponents/Component3/Component3.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './component3.scss';
+import './Component3.scss';
 
 const Component3 = () => {
