+++ src/renderer/components/app/app.scss
@@ -1,0 +1,20 @@
+.sideBar{
+    flex: 1;
+    padding: 0.3rem;
+    height: 100%;
+    background-color: lightgray;
+}
+
+.sideBar button {
+    margin-top: 0.1rem;
+}
+
+.content{
+    flex: 10;
+    padding: 2rem;
+}
+
+.wrapper{
+    display: flex;
+    height: 100%;
+}
