+++ src/renderer/components/app/app.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import './app.scss';
-import './router_components/description/description';
+import '../RouterComponents/Description/Description';
 import { Provider } from 'react-redux';
 import { configStore } from '../../../shared/configureStore';
-import UserLogin from './user_login/user_login';
-import CustomRouter from './custom_router/custom_router';
-import NavigationBar from './navigation_bar/navigation_bar';
+import UserLogin from '../UserLogin/UserLogin';
+import CustomRouter from '../CustomRouter/CustomRouter';
+import NavigationBar from '../NavigationBar/NavigationBar';
 
 const store = configStore('renderer');
 store.subscribe(() => console.log('action received in renderer'));
