--- src/renderer/components/app/custom_router/custom_router.scss
@@ -1,20 +1,0 @@
-.router {
-    display: flex;
-    height: 100%;
-}
-
-.sideBar{
-    flex: 1;
-    padding: 0.3rem;
-    height: 100%;
-    background-color: lightgray;
-}
-
-.sideBar button {
-    margin-top: 0.1rem;
-}
-
-.content{
-    flex: 10;
-    padding: 2rem;
-}
