--- src/renderer/components/app/router_components/description/description.scss
@@ -1,6 +1,0 @@
-.description {
-  color: blue;
-}
-.hello {
-  color: red;
-}
